Move conference home page to Jinja
