--- indico/modules/attachments/templates/mako_compat/attachments_button.html
@@ -1,3 +1,0 @@
-{% from 'attachments/_display.html' import render_attachments_button %}
-
-{{ render_attachments_button(files=attached_items.files, folders=attached_items.folders) }}
