--- indico/modules/attachments/templates/mako_compat/attachments_tree.html
@@ -1,17 +1,0 @@
-{% from 'attachments/_attachments.html' import render_attachments %}
-
-<div class="attachments-box {%- if can_edit %} highlighted-area{% endif %}">
-    <div class="attachments-box-header">
-        <h2 class="icon-attachment">{% trans %}Materials{% endtrans %}</h2>
-        {% if can_edit %}
-            <a href="#" id="manageAttachments-{{ event.id }}" class="right i-button icon-edit hide-if-locked"></a>
-        {% endif %}
-    </div>
-    {{ render_attachments(event.attached_items, management=false) }}
-</div>
-<script>
-    $('#manageAttachments-{{ event.id }}').on('click', function(evt) {
-        evt.preventDefault();
-        openAttachmentManager({{ event.locator|tojson }});
-    });
-</script>
