+++ indico/modules/events/views.py
@@ -22,7 +22,8 @@ from flask import render_template, request
 from sqlalchemy.orm import load_only
 
 from indico.legacy.webinterface.pages.base import WPDecorated, WPJinjaMixin
-from indico.legacy.webinterface.pages.conferences import (WConfMetadata, WPConferenceBase, WPrintPageFrame,
+from indico.legacy.webinterface.pages.conferences import (WConfMetadata, WPConferenceBase,
+                                                          WPConferenceDefaultDisplayBase, WPrintPageFrame,
                                                           render_event_footer, render_event_header)
 from indico.modules.admin.views import WPAdmin
 from indico.modules.events import Event
@@ -126,3 +127,13 @@ class WPAccessKey(WPJinjaMixin, WPDecorated):
 
     def _getBody(self, params):
         return self._getPageContent(params)
+
+
+class WPConferenceDisplay(WPConferenceDefaultDisplayBase):
+    menu_entry_name = 'overview'
+
+    def _getBody(self, params):
+        return render_template('events/display/conference.html', **self._kwargs)
+
+    def _getFooter(self):
+        return render_event_footer(self.event).encode('utf-8')
