+++ indico/web/flask/app.py
@@ -49,7 +49,7 @@ from indico.util import date_time as date_time_util
 from indico.util.i18n import _, babel, get_current_locale, gettext_context, ngettext_context
 from indico.util.mimetypes import icon_from_mimetype
 from indico.util.signals import values_from_signal
-from indico.util.string import alpha_enum, crc32, sanitize_html, slugify
+from indico.util.string import RichMarkup, alpha_enum, crc32, sanitize_html, slugify
 from indico.web.assets import (core_env, include_css_assets, include_js_assets, register_all_css, register_all_js,
                                register_theme_sass)
 from indico.web.flask.errors import errors_bp
@@ -207,6 +207,7 @@ def setup_jinja(app):
     app.add_template_filter(crc32)
     app.add_template_filter(bool)
     app.add_template_filter(lambda s: Markup(sanitize_html(s or '')), 'sanitize_html')
+    app.add_template_filter(RichMarkup, 'rich_markup')
     # Tests
     app.add_template_test(instanceof)  # only use this test if you really have to!
     app.add_template_test(subclassof)  # only use this test if you really have to!
